+++ mindspore/lite/src/runtime/kernel/arm/fp32/shape.cc
@@ -51,7 +51,7 @@ int ShapeCPUKernel::Run() {
     return RET_ERROR;
   }
 
-  for (int i = 0; i < in_tensor->shape().size(); i++) {
+  for (size_t i = 0; i < in_tensor->shape().size(); i++) {
     reinterpret_cast<int *>(out_tensor->Data())[i] = in_tensor->shape()[i];
   }
