+++ mindspore/lite/src/runtime/kernel/arm/fp32/unstack.cc
@@ -42,10 +42,10 @@ int UnstackCPUKernel::ReSize() {
   if (para->axis_ < 0) {
     para->axis_ += shape_size;
   }
-  for (size_t i = 0; i < shape_size; i++) {
-    if (i < para->axis_) {
+  for (size_t i = 0; i < static_cast<size_t>(shape_size); i++) {
+    if (static_cast<int>(i) < para->axis_) {
       para->pre_dims_ *= input->DimensionSize(i);
-    } else if (i > para->axis_) {
+    } else if (static_cast<int>(i) > para->axis_) {
       para->after_dims_ *= input->DimensionSize(i);
     } else {
       para->axis_dim_ = input->DimensionSize(i);
