+++ mindspore/lite/tools/converter/parser/tflite/tflite_arithmetic_parser.cc
@@ -127,7 +127,7 @@ STATUS TfliteDoubleInputOpParser::Parse(const std::unique_ptr<tflite::OperatorT>
   }
 
   // set input
-  for (int i = 0; i < tflite_op->inputs.size(); i++) {
+  for (size_t i = 0; i < tflite_op->inputs.size(); i++) {
     AddOpInput(op, tensors_id, tensors_format, tensors_id_map,
                tflite_op->inputs[i], tensors_id->size(), tflite_tensors.size(), schema::Format_NHWC);
   }
@@ -272,7 +272,7 @@ STATUS TfliteCompareOpParser::Parse(const std::unique_ptr<tflite::OperatorT> &tf
     op->primitive->value.value = attr.release();
   }
 
-  for (int i = 0; i < tflite_op->inputs.size(); i++) {
+  for (size_t i = 0; i < tflite_op->inputs.size(); i++) {
     AddOpInput(op, tensors_id, tensors_format, tensors_id_map,
                tflite_op->inputs[i], tensors_id->size(), tflite_tensors.size(), schema::Format_NHWC);
   }
