+++ mindspore/lite/tools/converter/parser/tflite/tflite_reshape_parser.cc
@@ -69,7 +69,7 @@ STATUS TfliteReshapeParser::Parse(const std::unique_ptr<tflite::OperatorT> &tfli
   op->primitive->value.type = schema::PrimitiveType_Reshape;
   op->primitive->value.value = attr.release();
 
-  for (int i = 0; i < tflite_op->inputs.size(); i++) {
+  for (size_t i = 0; i < tflite_op->inputs.size(); i++) {
     AddOpInput(op, tensors_id, tensors_format, tensors_id_map,
                tflite_op->inputs[i], tensors_id->size(), tflite_tensors.size(), schema::Format_NHWC);
   }
